+++ core/src/main/kotlin/no/elg/hex/Settings.kt
@@ -140,6 +140,7 @@ object Settings {
   )
 
   var enableDebugHUD by PreferenceDelegate(true, priority = 100_000, hideLevel = LOG_DEBUG)
+  var enableDebugFPSGraph by PreferenceDelegate(true, priority = 100_000, hideLevel = LOG_DEBUG)
 
   private const val DELETE_ALL_PROGRESS_STRING = "delete all"
   var deleteAllProgress by PreferenceDelegate(